This is a bug/problem fix update. No new features are being added.
I have tested each change. I beleive that no errors are being
introduced, and that the kind of changes being made have low
risk for introducing errors.

(1) ObjectViewer and ObjectReferenceInspector had errors on
zero length arrays due to an error in jdi. A workaround was coded.
This is an old bug, it was present in ObjectViewer the last public release.

(2) The inspectors of a project other than the first one opened were not
being loaded. This is fixed.

(3) In ObjectReferenceInspector, the initialization/update
of the object graph was moved into the JPanel's paint method. This
eliminates the (sometimes) noticeable delay when the panel is not currently
being used. Of course, if the user begins using the panel, they will observe
the delay at that time. They will not, however, be paying a penalty for
something they are not currently using, as it was before this update.

(4) MonashImageInspector is not general purpose, so it is a mistake to
distribute it in the installation lib "inspector" directory. Therefore, the
imageviewer example that works with MonashImageInspector was added so
that MonashImageInspector could be moved into the "(Inspector)"
subdirectory of imageviewer. MakeInspectors.bat was modified
to move the class files to the new directory. the UNIX make files
will need a change also. I did not do this.

(5) makeInpectors.bat, src/make.bat, and src/run.bat have been modified to
use environment variables for all absolute paths, and NOT explicitly set them.
One may now set the environment variables for their own installation, and
be able to use the bat files without modification.

(from subversion r757)