Second step in several refactorings necessary for exporting greenfoot projects to stand alone Java programs without including all of the greenfoot and bluej classes.
Refactored Actor and World. Moved most of the object transport (between VMs) functionality to the ObjectTracker. This avoids many dependencies from World and Actor to the rest of the Greenfoot classes.

Actor delegates a few method calls (that depends on environment) to a delegate interface (ActorDelegate). ActorDelegateIDE or ActorDelegateStandAlone are two implementations of this interface.

(from subversion r4799)