Made unit test recording work for more complex user interactions.
The major change is that it is now possible to get objects onto the bench from chained inspections (inspecting fields of objects). For instance, you can now inspect a method result and get one of the fields of the result onto the bench and still get valid unit test code.

(from subversion r5833)