Implemented a possible solution for bug #206: Give a warning/error if wrong case is used for filenames
This essentially makes Greenfoot case-sensitive on all systems.

If the wrong case is used for a filename on a system that is not case sensitive, an exception will be printed to the terminal. The exception is similar to the exception printed when a filename can not be found, but with more information about how to solve it. The exception will tell the user what the filename is supposed to be.

(from subversion r5989)