Performance enhancements for scenarios that uses a lot of Actors but only makes limited use of the collision checker.
The new class ColManager only adds an object to the "real" collision checker when a collision query has been made that involves the class of the object. For instance, if no collision check is ever made, the collision checker will never do any work, where previously it would keep all objects in the collision structure and update on changes to object bounds or position.

(from subversion r6066)