Removed Greep.setShip, and got rid of all the reflection.
To change Greeps you now have to change the name of the classes in Ship.createGreep().
I figured this was better than hardcording the class names to Greep1 and Greep2, since we would then have to rename the classes from the participants. Renaming classes would require renaming two files and editing the class name in the two files to match. Now we just have to edit one file.

As part of this change, Greep.getAuthorName is no longer a static method.

(from subversion r6088)