Sound support: major changes to the implementation of sound streams because previous strategy still made greenfoot hang on some systems.
Seemed to be timing issues related to opening and closing the same line many times within a short period of time. This new implementation relies on a minimum number of SourceDataLine method calls to avoid these strange issues, and so far it seems to work well. If it turns out that it doesn't work on some systems, I don't think there is much more we can do about it.

(from subversion r6278)