Fixed bug #229: sounds not located in scenario root directory on Greenfoot Gallery export.
The resource lookup now works a lot better when scenarios are published on the gallery. The problem was that the gallery always returns a valid URL even if you request a non-existing resource (because the gallery generates an error page). When the published scenario tries to look up a URL it now checks whether the returned URL is part of a jar, if it is not, it assumes that we got an incorrect URL back.

(from subversion r6380)