Corrected the auto-indent functionality in several ways.
Previously, when auto-indent was triggered, the caret would be set to the same offset in the file as before the indent.  This meant that if whitespace was removed, an exception could occur because we then attempted to set the caret beyond the end of the new file.  For the same reasons, the caret would jump around after an auto-indent in an annoying fashion.

The other thing corrected is the perfect indent calculation.  If the only change was that white-space was removed from the beginning of each line, indentation was wrongly considered perfect: for the perfect calculation, we only checked if the correct indent was already present at the beginning of the line, so 12 spaces being reduced to 8 was considered perfect because the line did begin with 8 spaces already.

(from subversion r7439)