Simplified the recording for save-the-world by no longer combining consecutive setLocation calls together (this used to be done to minimise the code generated for dragging an actor around).
Now if they drag the same actor twice in succession during recording, it will show up twice in the code.  But this means it will now be the same during recording and playback if they have overridden the setLocation method (previously, that might not have been the case: more calls would be made during recording than playback).

(from subversion r7765)