Renamed GreenfootRelauncher to GreenfootDebugHandler to reflect the other changes made:
Greenfoot now hooks into the debugger and listens for events relating to pausing the debugger or stepping.  Whenever the debugger pauses the VM at a point that is not somewhere inside the World's or Actor's act() methods (i.e. if World.act() or Actor.act() is not somewhere in the current call stack), the debugger runs the code further until the code is inside one of those methods and then stops.  This is kept hidden from the user, so they see the debugger as acting normally, but it will never pause in the Greenfoot simulation code.

Currently, the debugger can pause in a Greenfoot method that has been called from somewhere within an act() method (e.g. Greenfoot.isKeyDown(), etc).  There is also some debug output left in the class, as it's still a bit of a work in progress.

(from subversion r7851)