Allow BClass to represent classes that don't actually have a ClassTarget.
This is backwards compatible, unless, for some reason, an extension was relying
on getting an exception thrown when it called getBClass() on an object which
had a class not corresponding to a project class. This was, anyway,
undocumented behavior.

It makes the extensions API slightly more powerful, and is useful for Greenfoot.

(from subversion r8002)