Made parser errors (which are thrown as RuntimeException) encountered via the code pad be shown to the user rather than going uncaught and freezing the code pad.  This stops the problem with parser errors freezing the code pad.
This also seems to solve the NullPointerException problem when entering "true + 0", because NullPointerException is a RuntimeException.  I think that's more masking the problem than solving it, though, and makes me wonder if this solution is too broad.

Is there a reason parser errors don't have their own exception type?

(from subversion r8104)