Changed Greenfoot so that when the user clicks reset (and the debug handler notices via the special reset breakpoint, which now works correctly after the earlier commit), the object bench is cleared for the project.
This fixes the problem Michael found with the save the world functionality where a save after a reset (which previously did not clear the object bench -- only compile would do that) would end up removing the wrongly-named wombat.  Clearing the object bench means that all the naming works as intended.

(from subversion r8366)