Changed the ImageFrame for classes to have a button that displays a PopupMenu of the four actions, edit/duplicate/delete/new. Also some minor code clean up and documentation.
Added the PopupMenuButton that will set its appearance to the icon provided, and display the provided PopupMenu when clicked on.

Updated the english greenfoot labels to include tooltips for the first three actions to be consistent with the new image action. Also removed the unused labels for the removed buttons.

Also added the menu-button image to be displayed as the button on the frame.

(from subversion r8398)