Changed everything that synchronized on MachineLoaderThread to synchronize on the outer JdiDebugger instead.
There were various two-way and three-way deadlocks that could occur with threads holding one of the JdiDebugger and MachineLoaderThread monitors and wanting the other (e.g. one method was synchronized on JdiDebugger, but calling the synchronized method start() on the MachineLoaderThread, while another was calling getVM() which took the MachineLoaderThread monitor and then tried to get the JdiDebugger monitor -- there was also a 3-thread version!).  As far as I can see, making them all use the JdiDebugger monitor works and eliminates all these deadlocks.

This fixes a problem that would manifest as the whole program freezing up sometimes when you hit Terminate in the debugger.

(from subversion r8399)