Fixed two causes of deadlock when reset is pressed in Greenfoot.
removeObject was synchronized, but I don't see that it needed to be, and that caused a deadlock because while removeObject is executing (and waiting for the worker thread) there could be deadlocks with other methods that needed to synchronize on the VMReference.

The other one is that workerThread was using notify to wake up the next person waiting for the worker thread.  But some interactions involve waiting for the worker thread twice (once to set the action going, once to get the result after), and you could have a situation where one thread reserved the worker, set off and waited for the response, but the notify() after the task would wake up a second thread waiting to use the worker.  This second thread would see it was still reserved and go back to sleep, but the first would then never be woken up, resulting in deadlock (because the worker thread would never be finished with and become unreserved).

(from subversion r8401)