Further improve parser error reporting.
Most errors are due to an invalid token; in some cases the token actually found
is wrong and should be replaced, but in most cases something is just missing.
It makes more sense then not to highlight the "error" token but instead to display
an error bar before it.

(from subversion r8674)