Changed the behaviour of the automatic construction of a world following compilation.
Under the previous behaviour, if the last world could not be instantiated then the system would search all subclasses of world in the current project, and instantiate the first one it could successfully instantiate, so in practice you would end up with an arbitrary world being constructed if the last world wasn't compiled.  Also, any time that you hit compile on a single class (say an actor), and none of the world classes were compiled, you would get a dialog popping up telling you the world couldn't be instantiated (even though it had no error, it just didn't happen to be compiled).  That dialog now won't appear in that situation.

Now it is simple: if the last world can't be compiled, it makes no world -- effectively the same behaviour as before the multiple worlds functionality was added.

(from subversion r8834)