Fixed the check for whether there is a world during the simulation loop.
The previous code (including my previous attempted fix) was not enough.  It checked getWorld() != null, but getWorld() will return initialisingWorld when world is null, but what the simulation code wants to do is check if world is null.  So I've added a hasWorld() method that performs that check, which should stop the simulation thread doing anything when there is no world.

(from subversion r8844)