Fix a bug in the Counter class.
The Counter class has target (logically, the value of the counter), and value (which is what's currently displayed, which animates by counting up towards target).  Currently getValue() returns value, which looks right but is not what I'd expect the semantics to be: it should return target, otherwise it can return "out of date" values on its way to animating to the correct target.

I know this means building another RC, but I think it needs to be fixed.

(from subversion r9662)