Fixed the stderr buffering behaviour.
I hadn't realised, but the stderr part of the terminal always used to have unlimited buffering, and the recent changes to add stack-trace highlighting changed that to have the standard small buffer of the stdout portion (which could miss the beginning of long stack traces -- not good).

Now the stderr part has unlimited buffering again.

(from subversion r9752)