Added a String Greenfoot.ask(String prompt) method that allows you to ask for text input during a scenario's execution.  Pops up a panel at the base of the world with a text prompt and text field (and OK button) where you enter input and then execution continues.
The only difficulty seems to be if the user resets the world while waiting for input.  We have to interrupt the thread (in effect; we actually notify the monitor we know it is waiting on) and I think we then have to go back to user code before we can safely interrupt the simulation --  but to do that we need to return a value, so I'm returning the empty string for now.

(from subversion r12322)