Changed the show/hide behaviour of extends in the class header.
The rule is a bit more complex than discussed.  If you only show extends when the focus is in the name or extends slot, when you are in the (blank) extends slot, you get an odd behaviour moving right from the extends slot, where the extends slot would disappear and thus your target would move.

 There's now a latching behaviour, where if extends has been showing at all while your focus has been anywhere in the class header, the extends keeps showing until your focus leaves the class header completely.

 The only slightly odd behaviour remaining is if you have class Foo implements Bar, and you click the Bar then move left from the Bar, you end up in Foo, but the extends appears at that point, moving the place you left, which is a bit visually odd.  But it's the least disruptive solution so far, so that seems good enough.

(from subversion r16219)