The major part of the work in converting the object bench into JavaFX.  Currently unstyled but seems to be generally functional, pending some more testing.
There is a slight awkwardness where I created FX equivalents of all the methods to make menus with all the available methods to invoke -- but then realised that Greenfoot still uses the Swing version for when you right-click on actors or the world.  So currently there are both FX and Swing menu building methods, until we convert Greenfoot's main window into JavaFX.

The other awkwardness that we can't fix easily is that extensions are allowed to create Swing menu items to add to the invocation menu, which is now FX.  We do here what we already do elsewhere: we create FX "forwarding" menu items which reflect the contents of the underlying (non-visible) Swing items and forward events to the Swing items, without ever actually displaying the Swing item on screen.

(from subversion r16607)