Fixed the initial sizing of the class diagram and object bench panes.
This turned out to be a bigger task than expected.  The saving of the positions had been disabled in the move to FX because the pane size is known on the FX thread but the saving was being done on the Swing thread. Rather than use another thread-safe cache (basically: store the size in an AtomicInteger when it changes), I moved the saving to the FX thread (allowing us to remove a few existing caches) but it took a while to sort out the threading.

Also: JavaFX split panes are tricky.  There is a flag to stop the object bench proportionally resizing when the window resizes (what we want), but it must be set after we've loaded the size from the properties file because otherwise it prevents the size taking effect.

(from subversion r16695)