Added the JavaFX application to the object bench when using the special "Run JavaFX Application" menu option.
Previously, we ran the JavaFX application and it would show its window, but nothing would appear on the object bench.  If you then wanted to poke around your GUI objects using the object inspector, it was a bit hard because there was no "way in" to the object graph.  Since half the point of BlueJ is the object bench and all the inspection capabilities, this seemed wrong.

  So now, we add the Application instance itself to the object bench when you run it.  There is one additional tweak: usually, whenever we add an object to the bench, we ask for a name.  But that would get quite annoying quite quickly: you just want to test your FX application and you need to dismiss this extra (name entry) dialog every time. So we auto-assign the object name in this case to make the workflow shorter.

(from subversion r16822)