Fixed the behaviour for showing windows after compilation errors.
Previously, the behaviour was quite haphazard, and had several bugs/oddities: when you clicked compile and there was a class with an error, the editor would not show unless it had been shown before (in contrast to 3.1.7 behaviour where the editor would pop-up); when you tried to close a Stride editor it would focus back on PkgMgrFrame which triggered a compilation (more on that in next commit) which opened the Stride editor back up again, meaning you could never close a Stride editor which had an error.

Now the behaviour should be more uniform: if you explicitly compile from the PkgMgrFrame, all classes in which an error is detected will open (if several: as tabs in one editor window).  Might be a lot if many classes are in error; but then again, you just asked to compile all those classes with an error in them, so what did you expect?  (If you compile one selected class, only that one may pop up.)

If the compilation is only an error check then the window will not be shown, which prevents popping up windows at inappropriate times.  As part of this, softened several internal-compile events (e.g. on project load) into error-check-only in BlueJ, where we don't want the current Greenfoot behaviour of trying to recompile everything on project load.

(from subversion r16886)