Fixed a race hazard in Greenfoot caused by BlueJ 4 changes which would cause Greenfoot to deadlock when initialising the world on first load.
I got caught out by the fact that what is a cosmetic issue in BlueJ (the exact timing of adding objects to the bench) is a semantic requirement in Greenfoot.  So in BlueJ 4 we were thread-hopping from Swing to FX before adding an object to the bench, which didn't make any difference to the user.  But in Greenfoot this caused a deadlock because we expect that after calling ObjectBench.addObject, the difference will immediately be apparent to other methods called afterwards.  So now the two aspects are decoupled: there is an internal list in the object bench holding the true bench which is updated immediately, and we thread-hop to FX to update the GUI state afterwards.

(from subversion r16943)