Removed a lot of unneeded RMI classes.
So it turns out that there are several RMI classes that we aren't using, that were introduced in the initial RMI set back in 2004 but are now unused.  This commit removes these unused classes and methods.  The code compiles and seems to run fine without them. (I don't think we use any reflection on the RMI classes themselves, so I think the fact that it compiles indicates it's valid to remove them.).

(from subversion r16964)