Rejigged the behaviour of the status bar messages in the Java editor.
This is slightly different in BlueJ and Greenfoot because in Greenfoot we keep the classes resulting from an auto-compilation (therefore an auto-compilation really is an automatic proper compilation) whereas in BlueJ we throw them away (which can be thought of as an automatic error check which happens to use the compiler, rather than a compilation).  New behaviour is as follows.

In Greenfoot, whenever a compilation occurs, we first display the "File Saved" message, which is then replaced either by "Class compiled - no syntax errors" or "Error(s) found in class" depending on the result of the compilation.

In BlueJ, if you manually trigger the compilation with Ctrl/Cmd-K, you get the same behaviour as Greenfoot.  If it's an automatic error check, we first show "File Saved".  If there's errors we show "Error(s) found in class".  However, if there's no errors, we just leave up "File Saved".  (If we showed "Class compiled..." I think that would be confusing because the class diagram shows that the class needs compilation, and you can't invoke a method on the class until you compile.)

Fixes GREENFOOT-580

(from subversion r16996)