Fixed a bug in dependency tracking for Greenfoot.  I moved a bit more functionality from Package to PackageEditor than I should have, which meant dependencies weren't being recorded correctly in Greenfoot, which has no PackageEditor instance.  Now corrected, so dependencies will be tracked properly again in future, though any scenarios edited with trunk since 3.0.4 may have incomplete dependency information in the project file.
Symptoms of this bug were that if you had CrabWorld which created a Crab in its constructor (and thus CrabWorld depended on Crab), this dependency wasn't being recorded and thus edits to Crab did not result in CrabWorld being marked as needs-compile, which it should have been.  This meant if Crab had an error, you could still (incorrectly) instantiate CrabWorld, using the old Crab class on disk.

(from subversion r17004)