Removed the class files for the people example.
These files were the source of the "... modified in the future warning" which would often crop up during builds.  Because they get pulled out of version control, it's a toss up as to which of the .java and .class files is modified later and so you could get the wrong timestamps.  None of the other examples come pre-compiled, so I don't think this one needs to be (VCS tells me these files were originally added in 2003, and indeed they were still Java 1.4 class files)