Initial hacks to get a JavaFX version of MoeEditor.
This code will not pass the threadchecker at all, and lots of functionality is either removed (e.g. find/replace) or working weirdly (because MoeSyntaxDocument sourceDocument is still being used by many methods even though it's not what is showing on screen).  But it does compile (without thread checking) and run.  All parts of the code that have been modified or commented out with a temporary fix are commented "MOEFX" as a tag to return to and fix up later.