Initial work stopping MoeSyntaxDocument extending [Swing] Document class.
This class now wraps the FX document class, and has a mirror of the Swing Element API for ease of conversion.  Various other bits of code have been adjusted, or commented out (with MOEFX tag) to get it to compile and run again.  Still needs more work before it will start to function properly.