Tidy up some CSS color code, and enable user color schemes.
We enable user color schemes through a very simple mechanism: if they have a java-colors.css file in their BlueJ profile directory, it will override the built-in file of the same name.  So if they want a new set of colors, they can take a copy to their profile directory and edit it.