Add a link in displayed Javadoc to jump back to the BlueJ source for a method.
This does not modify the Javadoc on disk.  Instead it modifies the HTML in memory when it is shown in the Documentation view, setting up the listener for the link being clicked.  When the link is clicked, we switch back to the source view and move the caret to the relevant method.

(Figure we may as well get a new feature out of the switch to FX's WebView for viewing the documentation.)