Improved some aspects of the error marking.
The red underline placing is tricky because it must wrap horizontally (to extend to arbitrary lengths), but we don't want vertical wrapping, so we have to line up the image just right, in both standard resolution and Retina resolution.  This corrects it in standard, and hopefully doesn't break Retina.

Also added experimental display in the margin when the file is in an error state, trying to match the red error style on the class targets without being too distracting.