Eliminated scope background repaint flicker.
We were previously updating the backgrounds directly after every poll of the reparse queue but it seems this enabled intermediate states to become visible during the redraw phase.  Now we hold new backgrounds off-screen in a pending map until we are ready too apply all new backgrounds at a known point where they are consistent.  This seems to eliminate all the repaint flicker that I was seeing.