Reworked how we keep track of debugger threads for GUI display.
Previously, we used a Swing TreeModel which was in the bluej.debugger.jdi package.  This kept the debugger coupled more tightly to the GUI than is ideal, and also had an awkward syncer mechanism for thread hopping.  Also, a tree model seems like overkill; while it is correct that threads and thread groups form a tree, the cases where BlueJ programs use thread groups are surely very rare.

Now, we use a debugger thread listener interface (note: listener, not a model) to decouple the GUI from the debugger, and which keeps the thread hopping out of the debugger packages.  Because we need to track thread activity from the outset, regardless of whether the debugger is active, Project implements the interface, and stores the details which ExecControls can later use.