Fixed an issue where breakpoints set in JavaFX applications were not working.
We relaunch the VM before we run a JavaFX application, because of the one-FX-launch-per-VM-run rule that JavaFX imposes.  But we were sending the vm-created event (which then triggered the FX launch request from the server VM) before re-initing breakpoints, which meant that any breakpoints in e.g. the start method of a user's JavaFX Application subclass would probably be skipped.  This fix simply re-orders the two so that the breakpoints are re-inited before we send the event that will trigger the FX launch.