Changed PkgMgrFrame actions across to FX.
The old [Swing] "PkgMgrAction" is removed.  The [FX] "FXPkgMgrAction" has been renamed to PkgMgrAction, so many of the actions just continue to extend "PkgMgrAction", although it is not the FX version not the Swing version.

The team actions have also been rejigged.  They no longer extend FXAbstractAction, but instead are their own class.  Whenever you make a menu item or button you must pass the PkgMgrFrame reference in (because the actions are per-project, but the buttons and menu items affect e.g. the status text of the PkgMgrFrame they are in).  This avoids all the nasty passing of null and updating the PkgMgrFrame reference which the original design used.