Removed a lot of now-unnecessary runLaters in PkgMgrFrame.
PkgMgrFrame now passes the thread-checker (although not all the other classes do, so it may need to be tweaked again later).  Unfortunately, this commit alters a lot of indent levels and so touches a lot of lines which are not really altered.  If in future you get back to this commit during an annotate, you probably want to go further back to find the real change.