Added support for printing from PkgMgrFrame.
ProjectPrintDialog got folded into PrintDialog, so ProjectPrintDialog and SwingNodeDialog have now been removed.  There was a weird issue with printing the class diagram, so for now we are printing an image-copy of the diagram rather than the diagram itself.  Stride printing could do with some improvement (especially pagination) but we may live with that for this release.  All the Java classes print, and now without the dialog-popup-per-class behaviour that we had to live with during the mixed Swing/FX printing era.