Fixed the font CSS setting.
Stride was passing around the property with only the size, but Java was passing around the property including the -fx-font-size: prefix (and potentially, font family too).  This caused a conflict when they each used the same code completion class.  Now Stride has been changed to pass a complete piece of CSS to set the font size.