Fixed a problem which I believe was causing the editor freeze-up.
Sometimes when loading or scrolling the editor, it would freeze up.  Turns out this seems to be to do with a recently-added listener which was listening to visible-cells in the text view and then recalculating some styles, which may in turn affect the visible cells, leading to a later exception (which misleadingly had none of our code in it, so looked like it was nothing to do with us).  We now wrap the style recalculation in a runLater, which seems to fix the exception as far as I can tell (it only shows up sometimes).