Fixed how we blank out RichTextFX's built-in Ctrl/Cmd-Z/X/X/C/V bindings.
Previously, we consumed the keypress, which on my Mac still triggered the menu shortcut.  However, on my Windows PC (and, I believe, Michael's Mac) this prevented the menu accelerator being triggered.  Now fixed to use the ignore call instead of consume, which stops RichTextFX acting on the press, but still allows the menu accelerator or our custom shortcut to trigger.  Also now fixed so that if there is one shortcut for paste (e.g. F3) which ends up on the menu accelerator, but Ctrl-V is a second shortcut for paste, Ctrl-V still works fine (once, not twice or never) even though it's not on the accelerator.