Reset the state of compiling classes to needs-compile, even if it is only an error-check compile.
Previously, we only reset the state from error (red hashing) to needs-compile (grey stripes) if we kept the classes (e.g. it was a user-triggered full compile).  But this led to BLUEJ-992, where fixing an error led to the class target still showing red hashes until you did a full compile.  I think it's the correct behaviour to always switch back to grey slashes (then back again to red if there's an error), even if we're not keeping the classes.