Made the error message for compilation only show if compilation was explicitly requested.
You may be editing a class, and have an error that you know about (e.g. you're calling an unknown method, but you're working towards implementing that method).  Every time you change line in the source after editing, you trigger an error check compilation, and when it finds the error, it still tells you about it in the status bar (which will have been blanked because you've edited since last error).  This is irritating.

The new rule is: we only show the error message if you triggered a full compilation.  The "Errors: N" link is still shown after error check compilations, but not the message.