Fixed the padding around the bottom left status message in the Java editor.
There was actually reasonable padding set in em, but because the font size was set on the subcomponents, once you increased the font size, the em padding was still a multiplier on the default JavaFX font size, so it looked more crowded the larger you made the font size.  Now the font size is set on the status bar component itself so that the em-based sizes work properly.