Improved our performance on very large classes from unusable to slow.
It seems that RichTextFX is particularly poor for certain index and content queries on the document, because of the way it stores styled text.  This seems to make performance for calculating the scope highlights, where we do a lot of small queries of this kind, something like O(N^2) in the size of the class (or at least, worse than O(N)).  To combat this, we now cache the String content of the document and the position of line starts within that cache.  We use the cache whenever it is available, and invalidate it when the document changes.  While we are not very fast on large classes (I'm using PkgMgrFrame, 3500 lines), they do actually load after a delay rather than just seeming to hang the system forever.