Further improved the performance for long classes and those which don't fit on the screen.
Turns out, another cause of the slowness in the editor was that we were forever trying to recalculate the indent positions for nodes even when it couldn't possibly succeed (e.g. because the line wasn't on screen yet).  We now don't attempt calculation when it won't succeed, and we make better use of the cached space sizes to make node indent calculation more likely to succeed even when related lines are off-screen.  This should improve performance (especially loading) for all classes, but especially large classes.