Fixed one of the memory leaks in the Java editor.
We were sharing background images between source code lines.  Turns out, this means each line puts a listener on the image in case it changes (even though we know it never will).  This led to a situation where an on-screen line would keep live an image which might have listener-references to a potentially unbounded number of now-offscreen lines which should have been GCed, but wouldn't be because of the listeners.

We now sever the link by making copies of the image as needed.  Bit less memory efficient but the easiest way to solve the leak.  Also reduced the number of line backgrounds that we hold in each editor cache: we probably get better performance on lower-RAM machines by having less cached and just doing a little more redrawing during scrolling.