Fixed the other memory leak in the Java editor.
Each line number label needs to be notified if the line numbers on/off setting is changed.  We were adding a listener for each line label, with a weak reference to the label, for this purpose.  But while it let the labels get GCed, the listeners themselves were not GCed and thus built up a slow memory leak.  Now we have one listener, and a collection of a weak references which is cleared of old references when it is added to or read from.  This prevents a leak occurring relating to the line numbers when you scroll up and down a lot in a file (thus creating lots of new line number objects).