Removed a bluej.defs setting which caused the terminal to be tiny by default.
The terminal's old setting for width and height in number of characters was specified in bluej.defs.  This is meant to be no longer used (hence now removed), but it happened to use the same name as the new method for storing the terminal window's width and height (in pixels) in bluej.properties.  Because of the way our properties work, if you had no bluej.properties, we would unintentionally use the old setting from bluej.defs and make a tiny 80 pixel wide, 22 pixel high terminal window.