Fixed an issue where Stride was accidentally trying to calculate Java scope backgrounds behind the scenes.
Stride uses the parser for code completion, but is meant to have a null BlueJSyntaxView to prevent it trying to calculate scope backgrounds for the parsed code.  After some changes to MoeSyntaxDocument's constructors, it ended up having a non-null BlueJSyntaxView which led to all sorts of exceptions.  This sets the reference back to null as it should be, which prevents all the exceptions.