Fixed the scope drawing so that it uses the saved attributes for each paragraph as it should, rather than blanking them each time.  This was causing a problem where if you resized the window while you had a breakpoint or step marker showing, we would not repaint them because the attribute was only used for drawing if it had been set since last repaint.
Note: this commit merges a couple of similar methods in BlueJSyntaxView.  They were not identical (one got the attributes or a default empty set if missing, the other inserts the empty set if not found), but I believe the difference doesn't matter in our code (since missing is semantically equivalent to the empty set anyway).

Fixes BLUEJ-1028.

(It may also fix BLUEJ-1030 as error markers for a line are stored in the same way as step markers.)