Switched the generated_from_source_file_id item from being an attribute of a source file to being an attribute of source history.
This change makes sense because the generation state of a file can change over time (e.g. if you convert Stride to Java, the Java file changes from generated to not-generated) and we shouldn't destroy historical data which might be needed for analysis.  If we instead store it in the source histories table, it is preserved afterwards for analysis even if it changes again.