Fixed an issue where Ctrl-Space twice in Stride would not show all suggestions, as it is meant to.
This was caused by commit a8ee3be875909b874a31a7d18b1bb71e2403bfe7 (which fixed BLUEJ-883), which added a consuming key listener on a higher component in the hierarchy, thus preventing the keypress reaching the pre-existing listener.  The listeners have now been combined (on the higher element in the hierarchy).  This is not actually in any release, as that commit was since the last release, 4.1.0.