Changed the shown_error_indicators event to support sending a list of shown identifiers, rather than just one per master event.  Changed Java compilation to send all shown errors in a single event, rather than one event per error.
This will cut down greatly on the amount of events we need to send after a compilation: previously, we were sending one master event per compile error in each Java compilation, which was wasteful.  Stride errors are not changed by this commit, but they have the same issue so we can look to fix those next.