Fixed an issue with the debugger where the step/continue/halt buttons could get out of sync with the debugger.
This issue is two-fold.  The primary cause was that the thread details were only being updated properly if the thread was changed in the thread list, and otherwise it depended on whether the force-switch-thread flag was set.  Now, the details are updated properly.

The other issue was that the states of the suspended flag could get out of sync between the debugger and the GUI, because the GUI would look up the live value of the suspended flag, rather than the suspended flag at the time of the event.  So updates could arrive for resumed events with suspended set to true because the debugger had already stopped again.  We rectify that by taking a snapshot of the suspended state once for each event and processing that.

With these two changes, BLUEJ-1051 no longer occurs.