Fix an issue with showing HTML documentation for a class, where the WebView component could get into a state where it would never show the documentation for a new class.
Coming back to this with fresh eyes, the problem was a call to refreshHtmlDisplay() before the docs existed, which attempted [and failed] a load, and no amount of reloading afterwards will fix the problem.  To avoid this, we don't attempt to load the docs page until the file exists.  Plus a little rearranging to avoid attempting the load if we're about to regenerate anyway.

(Note: the check inside refreshHtmlDisplay is theoretically unneeded due to fixing the callers, but I'd prefer to leave it in than add a pre-condition to the refreshHtmlDisplay about the docs must exist, which future callers may overlook.)